+++ packages/platform-android/src/lib/commands/buildAndroid/buildAndroid.ts
@@ -1,10 +1,8 @@
 import type { AndroidProjectConfig } from '@react-native-community/cli-types';
-import type { RemoteBuildCache } from '@rnef/tools';
 import {
   colorLink,
   type FingerprintSources,
   formatArtifactName,
-  getBinaryPath,
   logger,
   outro,
   parseArgs,
@@ -20,14 +18,12 @@ export interface BuildFlags {
   activeArchOnly?: boolean;
   tasks?: Array<string>;
   extraParams?: Array<string>;
-  local?: boolean;
 }
 
 export async function buildAndroid(
   androidProject: AndroidProjectConfig,
   args: BuildFlags,
   projectRoot: string,
-  remoteCacheProvider: null | (() => RemoteBuildCache) | undefined,
   fingerprintOptions: FingerprintSources
 ) {
   normalizeArgs(args);
@@ -40,19 +36,9 @@ export async function buildAndroid(
     root: projectRoot,
     fingerprintOptions,
   });
-  const binaryPath = await getBinaryPath({
-    artifactName,
-    localFlag: args.local,
-    remoteCacheProvider,
-  });
-  if (binaryPath) {
-    //
-  } else {
-    await runGradle({ tasks, androidProject, args, artifactName });
-  }
+  await runGradle({ tasks, androidProject, args, artifactName });
 
-  const outputFilePath =
-    binaryPath ?? (await findOutputFile(androidProject, tasks));
+  const outputFilePath = await findOutputFile(androidProject, tasks);
 
   if (outputFilePath) {
     logger.log(
@@ -98,9 +84,5 @@ export const options = [
     name: '--extra-params <string>',
     description: 'Custom params passed to gradle build command',
     parse: parseArgs,
-  },
-  {
-    name: '--local',
-    description: 'Force local build with Gradle wrapper.',
   },
 ];
