+++ packages/platform-android/src/lib/commands/runAndroid/runAndroid.ts
@@ -7,7 +7,9 @@ import type {
 import type { FingerprintSources, RemoteBuildCache } from '@rnef/tools';
 import {
   color,
+  fetchCachedBuild,
   formatArtifactName,
+  getLocalBuildCacheBinaryPath,
   intro,
   isInteractive,
   logger,
@@ -16,7 +18,6 @@ import {
   RnefError,
   spinner,
 } from '@rnef/tools';
-import { getBinaryPath } from '@rnef/tools';
 import type { BuildFlags } from '../buildAndroid/buildAndroid.js';
 import { options } from '../buildAndroid/buildAndroid.js';
 import { runGradle } from '../runGradle.js';
@@ -67,12 +68,35 @@ export async function runAndroid(
     root: projectRoot,
     fingerprintOptions,
   });
-  const binaryPath = await getBinaryPath({
-    artifactName,
-    binaryPathFlag: args.binaryPath,
-    localFlag: args.local,
-    remoteCacheProvider,
-  });
+  // 1. First check if the binary path is provided
+  let binaryPath = args.binaryPath;
+
+  // 2. If not, check if the local build is requested
+  if (!binaryPath && !args.local) {
+    binaryPath = getLocalBuildCacheBinaryPath(artifactName);
+  }
+
+  // 3. If not, check if the remote cache is requested
+  if (!binaryPath && !args.local) {
+    try {
+      const cachedBuild = await fetchCachedBuild({
+        artifactName,
+        remoteCacheProvider,
+      });
+      if (cachedBuild) {
+        binaryPath = cachedBuild.binaryPath;
+      }
+    } catch (error) {
+      const message = (error as RnefError).message;
+      const cause = (error as RnefError).cause;
+      logger.warn(
+        `Failed to fetch cached build for ${artifactName}: \n${message}`,
+        cause ? `\nCause: ${cause.toString()}` : ''
+      );
+      logger.debug('Remote cache failure error:', error);
+      logger.info('Continuing with local build');
+    }
+  }
 
   if (device) {
     if (!(await getDevices()).find((d) => d === device.deviceId)) {
@@ -273,5 +297,9 @@ export const runOptions = [
   {
     name: '--user <number>',
     description: 'Id of the User Profile you want to install the app on.',
+  },
+  {
+    name: '--local',
+    description: 'Force local build with Gradle wrapper.',
   },
 ];
