+++ packages/platform-apple-helpers/src/lib/commands/run/runOptions.ts
@@ -22,6 +22,10 @@ export const getRunOptions = ({ platformName }: BuilderCommand) => {
         'Path relative to project root where pre-built .app binary lives.',
     },
     {
+      name: '--local',
+      description: 'Force local build with xcodebuild.',
+    },
+    {
       name: '--device <string>',
       description:
         'Explicitly set the device or simulator to use by name or by UDID.',
