+++ packages/platform-ios/src/lib/platformIOS.ts
@@ -35,7 +35,6 @@ export const platformIOS =
           projectRoot,
           reactNativePath: api.getReactNativePath(),
           fingerprintOptions: api.getFingerprintOptions(),
-          remoteCacheProvider: await api.getRemoteCacheProvider(),
         });
         outro('Success 🎉.');
       },
